refactor(services): move detection-count helper to services/
Per review feedback on PR #559: get_detection_counts_by_sequence_ids
is an aggregation query, not a CRUD/route helper, so it belongs in
services/ where it can be reused outside of endpoint modules.
